update to run many model tests
update to run many model tests

includes tools for in utils and image_tools for changing images and labels to use multi class info and multiple time points for different color channels.

also added split_h5_loop_segments to image_tools to randomly sample H5 images using groups based on frame nums. this is so we can see the temporal information when looking at classification later. like if a touch was missed in the middle of a touch.